+++ CHANGELOG.md
@@ -34,7 +34,9 @@ Using the following categories, list your changes in this order:
 
 ## [Unreleased]
 
--   Nothing (yet)!
+### Added
+
+-   Warning W018 (`Suspicious position of 'reactpy_django' in INSTALLED_APPS`) has been added.
 
 ## [3.5.0] - 2023-08-26